socialnet_mono/requirements.txt
Django==5.2.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
django-redis>=5.4
celery>=5.3
django-celery-results>=2.5
gunicorn>=21.2
boto3>=1.34
django-storages>=1.14
Pillow>=10.3
minio>=7.2
